+++ CJBCheatsMenu/Framework/Cheats/FarmAndFishing/InfiniteHayCheat.cs
@@ -44,9 +44,19 @@ namespace CJBCheatsMenu.Framework.Cheats.FarmAndFishing
             if (!e.IsOneSecond || !Context.IsWorldReady)
                 return;
 
-            Farm? farm = Game1.getFarm();
-            if (farm != null)
-                farm.piecesOfHay.Value = Utility.numSilos() * 240;
+            Utility.ForEachLocation(
+                location =>
+                {
+                    if (location.buildings.Count > 0)
+                    {
+                        int addHay = location.GetHayCapacity() - location.piecesOfHay;
+                        location.tryToAddHay(addHay);
+                    }
+
+                    return true;
+                },
+                includeInteriors: false
+            );
         }
     }
 }