+++ CJBCheatsMenu/Framework/GameI18n.cs
@@ -1,7 +1,6 @@
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 using StardewValley;
+using StardewValley.ItemTypeDefinitions;
 
 namespace CJBCheatsMenu.Framework
 {
@@ -13,47 +12,31 @@ namespace CJBCheatsMenu.Framework
         ** Public methods
         *********/
         /// <summary>Get the translated name for a big craftable object.</summary>
-        /// <param name="id">The big craftable ID.</param>
-        public static string GetBigCraftableName(int id)
+        /// <param name="id">The unqualified big craftable ID.</param>
+        public static string GetBigCraftableName(string id)
         {
-            if (Game1.bigCraftablesInformation == null)
+            if (Game1.bigCraftableData is null)
                 return "(missing translation: game hasn't loaded big craftable data yet)";
-            if (!Game1.bigCraftablesInformation.TryGetValue(id, out string? value))
+
+            ParsedItemData? data = ItemRegistry.GetData(ItemRegistry.type_bigCraftable + id);
+            if (data is null)
                 return $"(missing translation: no big craftable #{id})";
 
-            return value.Split('/').Last();
-        }
-
-        /// <summary>Get the translated name for a building.</summary>
-        /// <param name="id">The building ID.</param>
-        public static string GetBuildingName(string id)
-        {
-            var data = Game1.content.Load<Dictionary<string, string>>("Data\\Blueprints");
-
-            if (!data.TryGetValue(id, out string? rawData))
-                return $"(missing translation: no building with ID '{id}')";
-
-            string[] fields = rawData.Split('/');
-            if (fields.Length <= 8)
-                return $"(missing translation: building with ID '{id}' has invalid data)";
-
-            return fields[8];
+            return data.DisplayName;
         }
 
         /// <summary>Get the translated name for an object.</summary>
-        /// <param name="id">The object ID.</param>
-        public static string GetObjectName(int id)
+        /// <param name="id">The unqualified object ID.</param>
+        public static string GetObjectName(string id)
         {
-            if (Game1.objectInformation == null)
+            if (Game1.objectData is null)
                 return "(missing translation: game hasn't loaded object data yet)";
-            if (!Game1.objectInformation.TryGetValue(id, out string? value))
+
+            ParsedItemData? data = ItemRegistry.GetData(ItemRegistry.type_object + id);
+            if (data is null)
                 return $"(missing translation: no object #{id})";
 
-            string[] parts = value.Split('/');
-            if (parts.Length < 5)
-                return $"(missing translation: object #{id} has an invalid format)";
-
-            return parts[4];
+            return data.DisplayName;
         }
     }
 }