+++ CJBItemSpawner/Framework/ItemMenu.cs
@@ -830,7 +830,10 @@ namespace CJBItemSpawner.Framework
             {
                 ItemSort.Price => items.OrderByDescending(p => this.GetSellPrice(p.Item) ?? -1).ThenBy(p => p.Item.DisplayName),
                 ItemSort.Type => items.OrderBy(p => p.Item.Category),
-                ItemSort.ID => items.OrderBy(p => p.Item.ParentSheetIndex),
+                ItemSort.ID => items.OrderBy(
+                    p => int.TryParse(p.Item.ItemId, out int index) ? index.ToString().PadLeft(10, '0') : p.Item.ItemId,
+                    StringComparer.OrdinalIgnoreCase
+                ),
                 _ => items.OrderBy(p => p.Item.DisplayName)
             };
 
@@ -910,8 +913,8 @@ namespace CJBItemSpawner.Framework
             if (item is null)
                 return;
 
-            if (item is SObject obj && Game1.player.specialItems.Contains(obj.ParentSheetIndex))
-                Game1.player.specialItems.Remove(obj.ParentSheetIndex);
+            if (item is SObject obj && Game1.player.specialItems.Contains(obj.ItemId))
+                Game1.player.specialItems.Remove(obj.ItemId);
             if (this.ReclaimPriceInTrashCan)
             {
                 int price = Utility.getTrashReclamationPrice(item, Game1.player);